+++ CHANGES.md
@@ -1,5 +1,13 @@
 ### Unreleased
 
+#### Add tactic admit (2021-03-12)
+
+- rename command `admit` into `admitted`
+- `admitted`: admit the initial goal instead of the remaining goals (when the proof is an opaque definition)
+- move code on `admit` from `command.ml` to `tactic.ml`
+- add tactic `admit` (fix #380)
+  As a consequence, tactics can change the signature state now.
+
 #### Improvements in type inference, unification and printing (2021-03-11)
 
 - improve type inference and unification